Upravljanje artiklima
Kompletna izrada funkcionalnosti Upravljanje artiklima, kreiran DLL za generiranje barkoda (koristit će se i za generiranje aktivacijskog koda kod registracije), izvršeno nekoliko promjena nad bazom podataka a samim time mijenjan entity model, zbog čega je bilo potrebno prilagoditi i klase drugih funkcionalnosti (promjena tipa podataka za ID artikla), kreirano/izmijenjeno nekoliko parcijalnih klasa entity modela zbog "customizacija"

+++ Software/MiniStore/ArtiklPovratnica.cs
@@ -14,7 +14,7 @@ namespace MiniStore
     
     public partial class ArtiklPovratnica
     {
-        public int artiklId { get; set; }
+        public long artiklId { get; set; }
         public int povratnicaId { get; set; }
         public Nullable<int> kolicina { get; set; }
         public Nullable<decimal> cijena { get; set; }
