+++ Software/MiniStore/ArtiklPrimka.cs
@@ -14,7 +14,7 @@ namespace MiniStore
     
     public partial class ArtiklPrimka
     {
-        public int artiklId { get; set; }
+        public long artiklId { get; set; }
         public int primkaId { get; set; }
         public Nullable<int> dobavljacId { get; set; }
         public Nullable<int> skladisteId { get; set; }
